+++ spotlight-client/src/VizNotes/UnknownsNote.tsx
@@ -15,91 +15,9 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { Link } from "@reach/router";
 import { Button } from "@recidiviz/design-system";
-import { format } from "date-fns";
-import React, { useState } from "react";
-import styled from "styled-components/macro";
-import { ValuesType } from "utility-types";
-import { UnknownCounts, Unknowns } from "../contentModels/types";
-import { getDemographicViewLabel } from "../demographics";
-import { DemographicFieldKeyList } from "../metricsApi";
+import React from "react";
 import { AutoHeightTransition } from "../UiLibrary";
-import { formatAsNumber } from "../utils";
-
-const COLLAPSED_LIMIT = 3;
-
-const ExpandNoteButton = styled(Button).attrs({ kind: "link" })`
-  font-size: inherit;
-  font-weight: inherit;
-`;
-
-function formatUnknownCounts(unknowns: UnknownCounts) {
-  const parts: string[] = [];
-
-  DemographicFieldKeyList.forEach((key) => {
-    const value = unknowns[key];
-    if (!value) return;
-
-    parts.push(
-      `${getDemographicViewLabel(key).toLowerCase()} (${formatAsNumber(value)})`
-    );
-  });
-
-  return parts.join(", ");
-}
-
-function useFormattedUnknowns(unknowns: Unknowns) {
-  // if the list is short this will have no effect; if it's long we want to start collapsed
-  const [isCollapsed, setIsCollapsed] = useState(true);
-
-  if (Array.isArray(unknowns)) {
-    const isCollapsible = unknowns.length > COLLAPSED_LIMIT;
-
-    // Typescript freaks out over unions of array types,
-    // but this assertion is functionally identical
-    let unknownsToShow = unknowns as ValuesType<typeof unknowns>[];
-    let overflowCount = 0;
-
-    if (isCollapsed) {
-      unknownsToShow = unknownsToShow.slice(0, COLLAPSED_LIMIT);
-      overflowCount = unknowns.length - unknownsToShow.length;
-    }
-
-    const formattedUnknowns = unknownsToShow
-      .map((entry) => {
-        const formattedCounts = formatUnknownCounts(entry.unknowns);
-        let label: string;
-
-        if ("date" in entry) {
-          label = format(entry.date, "MMM d y");
-        } else {
-          label = `the ${entry.cohort} cohort`;
-        }
-
-        return `${formattedCounts} for ${label}`;
-      })
-      .join("; ");
-
-    return (
-      <>
-        {formattedUnknowns}
-        {isCollapsible && (
-          <>
-            {isCollapsed ? "; " : " "}
-            <ExpandNoteButton onClick={() => setIsCollapsed(!isCollapsed)}>
-              {isCollapsed
-                ? ` + ${overflowCount} more`
-                : "(Hide extended list)"}
-            </ExpandNoteButton>
-          </>
-        )}
-      </>
-    );
-  }
-
-  return formatUnknownCounts(unknowns);
-}
 
 type UnknownsNoteProps = {
   download: () => Promise<void>;
@@ -113,9 +31,9 @@ export const UnknownsNote = ({ download }: UnknownsNoteProps): JSX.Element => {
       but are excluded from demographic breakdown views, counting towards a
       maximum difference of X% between the total and the sum of broken down
       figures. For more information,{" "}
-      <Link to="" onClick={download}>
+      <Button kind="link" onClick={download}>
         download
-      </Link>{" "}
+      </Button>{" "}
       the data and view the README.txt.
     </AutoHeightTransition>
   );
