+++ spotlight-client/src/contentModels/Metric.ts
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
+import { format } from "date-fns";
 import {
   action,
   computed,
@@ -34,6 +35,7 @@ import {
   DemographicCategories,
   DemographicView,
   getDemographicCategoriesForView,
+  getDemographicViewLabel,
 } from "../demographics";
 import {
   RawMetricData,
@@ -41,9 +43,33 @@ import {
   LocalityFields,
   SupervisionSuccessRateMonthlyRecord,
   fetchAndTransformMetric,
+  DemographicFieldKeyList,
 } from "../metricsApi";
+import { formatAsNumber } from "../utils";
 import downloadData from "./downloadData";
-import { MetricRecord, Hydratable } from "./types";
+import {
+  MetricRecord,
+  Hydratable,
+  Unknowns,
+  UnknownCounts,
+  UnknownByDate,
+  UnknownByCohort,
+} from "./types";
+
+function formatUnknownCounts(unknowns: UnknownCounts) {
+  const parts: string[] = [];
+
+  DemographicFieldKeyList.forEach((key) => {
+    const value = unknowns[key];
+    if (!value) return;
+
+    parts.push(
+      `${getDemographicViewLabel(key).toLowerCase()} (${formatAsNumber(value)})`
+    );
+  });
+
+  return parts.join(", ");
+}
 
 export type BaseMetricConstructorOptions<RecordFormat extends MetricRecord> = {
   id: MetricTypeId;
@@ -135,12 +161,11 @@ export default abstract class Metric<RecordFormat extends MetricRecord>
       hydrate: action,
       isLoading: observable,
       records: computed,
-      readme: computed,
     });
 
     // initialize metadata
     this.name = name;
-    this.methodology = methodology;
+    this.methodology = methodology.replace(/\s+/g, " ");
     this.id = id;
 
     // initialize data fetching
@@ -195,8 +220,39 @@ export default abstract class Metric<RecordFormat extends MetricRecord>
     return this.allRecords;
   }
 
-  get readme(): string {
-    return this.methodology;
+  get unknowns(): Unknowns | undefined {
+    return undefined;
+  }
+
+  generateReadme(): string {
+    const unknowns = this.unknowns;
+    if (!unknowns) {
+      return this.methodology;
+    }
+    let formattedUnknowns: string;
+    if (Array.isArray(unknowns)) {
+      formattedUnknowns = (unknowns as (UnknownByDate | UnknownByCohort)[])
+        .map((entry) => {
+          const formattedCounts = formatUnknownCounts(entry.unknowns);
+          let label: string;
+
+          if ("date" in entry) {
+            label = format(entry.date, "MMM d y");
+          } else {
+            label = `the ${entry.cohort} cohort`;
+          }
+
+          return `${formattedCounts} for ${label}`;
+        })
+        .join("; ");
+    } else {
+      formattedUnknowns = formatUnknownCounts(unknowns);
+    }
+    return (
+      this.methodology +
+      `\n\nVISUALIZATION FOOTNOTES:\n\nValues that count towards the total but are excluded from demographic breakdown views: ` +
+      formattedUnknowns
+    );
   }
 
   /**
@@ -211,7 +267,7 @@ export default abstract class Metric<RecordFormat extends MetricRecord>
     return runInAction(() =>
       downloadData({
         archiveName: `${this.tenantId} ${this.id} data`,
-        readmeContents: this.readme,
+        readmeContents: this.generateReadme(),
         // allRecords won't be undefined because we just awaited it
         dataFiles: [{ name: "data", data: this.allRecords as RecordFormat[] }],
       })
